+++ backend/src/main/java/com/example/backend/domain/member/entity/Member.java
@@ -1,7 +1,7 @@
 package com.example.backend.domain.member.entity;
 
 import com.example.backend.domain.member.dto.MemberModifyRequestDto;
-import com.example.backend.global.auth.dto.KakaoUserInfoResponseDto;
+import com.example.backend.global.auth.kakao.dto.KakaoUserInfoResponseDto;
 import com.example.backend.global.base.BaseEntity;
 
 import jakarta.persistence.Column;
