+++ backend/src/main/java/com/example/backend/global/auth/jwt/MemberAuthFilter.java
@@ -11,10 +11,10 @@ import com.example.backend.domain.member.dto.MemberTokenReissueDto;
 import com.example.backend.domain.member.exception.MemberException;
 import com.example.backend.global.auth.exception.AuthErrorCode;
 import com.example.backend.global.auth.exception.AuthException;
+import com.example.backend.global.auth.kakao.service.KakaoAuthService;
 import com.example.backend.global.auth.model.CustomUserDetails;
 import com.example.backend.global.auth.service.CookieService;
 import com.example.backend.global.auth.service.CustomUserDetailService;
-import com.example.backend.global.auth.service.KakaoAuthService;
 import com.example.backend.global.auth.util.JwtUtil;
 import com.example.backend.global.auth.util.TokenProvider;
 import com.example.backend.global.response.ErrorResponse;
