+++ backend/src/main/java/com/example/backend/global/auth/kakao/dto/KakaoTokenResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.backend.global.auth.dto;
+package com.example.backend.global.auth.kakao.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
