+++ backend/src/main/java/com/example/backend/global/auth/kakao/service/KakaoAuthService.java
@@ -1,4 +1,4 @@
-package com.example.backend.global.auth.service;
+package com.example.backend.global.auth.kakao.service;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatusCode;
@@ -11,12 +11,13 @@ import org.springframework.web.reactive.function.client.WebClient;
 import com.example.backend.domain.member.dto.MemberTokenReissueDto;
 import com.example.backend.domain.member.entity.Member;
 import com.example.backend.domain.member.service.MemberService;
-import com.example.backend.global.auth.dto.KakaoTokenResponseDto;
-import com.example.backend.global.auth.dto.KakaoUserInfoResponseDto;
-import com.example.backend.global.auth.dto.LoginResponseDto;
-import com.example.backend.global.auth.exception.KakaoAuthErrorCode;
-import com.example.backend.global.auth.exception.KakaoAuthException;
-import com.example.backend.global.auth.util.KakaoAuthUtil;
+import com.example.backend.global.auth.kakao.dto.KakaoTokenResponseDto;
+import com.example.backend.global.auth.kakao.dto.KakaoUserInfoResponseDto;
+import com.example.backend.global.auth.kakao.dto.LoginResponseDto;
+import com.example.backend.global.auth.kakao.exception.KakaoAuthErrorCode;
+import com.example.backend.global.auth.kakao.exception.KakaoAuthException;
+import com.example.backend.global.auth.kakao.util.KakaoAuthUtil;
+import com.example.backend.global.auth.service.CookieService;
 import com.example.backend.global.auth.util.TokenProvider;
 
 import jakarta.servlet.http.HttpServletResponse;
