+++ backend/src/main/java/com/example/backend/global/auth/kakao/util/KakaoAuthUtil.java
@@ -1,4 +1,4 @@
-package com.example.backend.global.auth.util;
+package com.example.backend.global.auth.kakao.util;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
